Move external forms from workshops to extforms app
This fixes #816
