+++ workshops/views.py
@@ -23,7 +23,6 @@ from django.db.models import Count, Q, F, Model, ProtectedError, Sum
 from django.http import Http404, HttpResponse, JsonResponse
 from django.http import HttpResponseBadRequest, HttpResponseRedirect
 from django.shortcuts import redirect, render, get_object_or_404
-from django.template.loader import get_template
 from django.utils.http import is_safe_url
 from django.views.generic import ListView, DetailView, TemplateView, DeleteView
 from django.views.generic.edit import (
@@ -58,9 +57,6 @@ from workshops.forms import (
     PersonsSelectionForm,
     PersonTaskForm,
     OrganizationForm,
-    SWCEventRequestForm,
-    DCEventRequestForm,
-    ProfileUpdateRequestForm,
     PersonLookupForm,
     SimpleTodoForm,
     BootstrapHelper,
@@ -72,14 +68,11 @@ from workshops.forms import (
     EventsMergeForm,
     InvoiceRequestForm,
     InvoiceRequestUpdateForm,
-    EventSubmitForm,
     EventSubmitFormNoCaptcha,
     PersonsMergeForm,
     PersonCreateForm,
     SponsorshipForm,
-    TrainingRequestForm,
     AutoUpdateProfileForm,
-    DCSelfOrganizedEventRequestForm,
     DCSelfOrganizedEventRequestFormNoCaptcha,
     TrainingProgressForm,
     BulkAddTrainingProgressForm,
@@ -2348,88 +2341,6 @@ def object_changes(request, revision_id):
 # ------------------------------------------------------------
 
 
-class SWCEventRequest(LoginNotRequiredMixin, EmailSendMixin, CreateViewContext):
-    model = EventRequest
-    form_class = SWCEventRequestForm
-    page_title = 'Request a Software Carpentry Workshop'
-    template_name = 'forms/workshop_swc_request.html'
-    success_url = reverse_lazy('swc_workshop_request_confirm')
-    email_fail_silently = False
-    email_kwargs = {
-        'to': settings.REQUEST_NOTIFICATIONS_RECIPIENTS,
-        'reply_to': None,
-    }
-
-    def get_success_message(self, *args, **kwargs):
-        """Don't display a success message."""
-        return ''
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = self.page_title
-        return context
-
-    def get_subject(self):
-        subject = (
-            '[{tag}] New workshop request: {affiliation}, {country}'
-        ).format(
-            tag=self.object.workshop_type.upper(),
-            country=self.object.country.name,
-            affiliation=self.object.affiliation,
-        )
-        return subject
-
-    def get_body(self):
-        link = self.object.get_absolute_url()
-        link_domain = settings.SITE_URL
-
-        body_txt = get_template(
-            'mailing/eventrequest.txt'
-        ).render({
-            'object': self.object,
-            'link': link,
-            'link_domain': link_domain,
-        })
-
-        body_html = get_template(
-            'mailing/eventrequest.html'
-        ).render({
-            'object': self.object,
-            'link': link,
-            'link_domain': link_domain,
-        })
-        return body_txt, body_html
-
-    def form_valid(self, form):
-        """Send email to admins if the form is valid."""
-        data = form.cleaned_data
-        self.email_kwargs['reply_to'] = (data['email'], )
-        result = super().form_valid(form)
-        return result
-
-
-class SWCEventRequestConfirm(LoginNotRequiredMixin, TemplateView):
-    """Display confirmation of received workshop request."""
-    template_name = 'forms/workshop_swc_request_confirm.html'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = 'Thank you for requesting a workshop'
-        return context
-
-
-class DCEventRequest(SWCEventRequest):
-    form_class = DCEventRequestForm
-    page_title = 'Request a Data Carpentry Workshop'
-    template_name = 'forms/workshop_dc_request.html'
-    success_url = reverse_lazy('dc_workshop_request_confirm')
-
-
-class DCEventRequestConfirm(SWCEventRequestConfirm):
-    """Display confirmation of received workshop request."""
-    template_name = 'forms/workshop_dc_request_confirm.html'
-
-
 @admin_required
 def all_eventrequests(request):
     """List all event requests."""
@@ -2526,40 +2437,6 @@ def eventrequest_assign(request, request_id, person_id=None):
     event_req = get_object_or_404(EventRequest, pk=request_id)
     assign(request, event_req, person_id)
     return redirect(reverse('eventrequest_details', args=[event_req.pk]))
-
-
-@login_not_required
-def profileupdaterequest_create(request):
-    """
-    Profile update request form. Accessible to all users (no login required).
-
-    This one is used when instructors want to change their information.
-    """
-    form = ProfileUpdateRequestForm()
-    page_title = 'Update Instructor Profile'
-
-    if request.method == 'POST':
-        form = ProfileUpdateRequestForm(request.POST)
-
-        if form.is_valid():
-            form.save()
-
-            # TODO: email notification?
-
-            context = {
-                'title': 'Thank you for updating your instructor profile',
-            }
-            return render(request,
-                          'forms/profileupdate_confirm.html',
-                          context)
-        else:
-            messages.error(request, 'Fix errors below.')
-
-    context = {
-        'title': page_title,
-        'form': form,
-    }
-    return render(request, 'forms/profileupdate.html', context)
 
 
 class AllProfileUpdateRequests(OnlyForAdminsMixin, ListView):
@@ -2749,61 +2626,6 @@ def profileupdaterequest_accept(request, request_id, person_id=None):
     return redirect(person.get_absolute_url())
 
 
-# This form is disabled as per @maneesha's request
-# class EventSubmission(LoginNotRequiredMixin, EmailSendMixin,
-#                       CreateViewContext):
-class EventSubmission(LoginNotRequiredMixin, TemplateView):
-    """Display form for submitting existing workshops."""
-    model = EventSubmissionModel
-    form_class = EventSubmitForm
-    template_name = 'forms/event_submit.html'
-    success_url = reverse_lazy('event_submission_confirm')
-    email_fail_silently = False
-    email_kwargs = {
-        'to': settings.REQUEST_NOTIFICATIONS_RECIPIENTS,
-    }
-
-    def get_success_message(self, *args, **kwargs):
-        """Don't display a success message."""
-        return ''
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = 'Tell us about your workshop'
-        return context
-
-    def get_subject(self):
-        return ('New workshop submission from {}'
-                .format(self.object.contact_name))
-
-    def get_body(self):
-        link = self.object.get_absolute_url()
-        link_domain = settings.SITE_URL
-        body_txt = get_template('mailing/event_submission.txt') \
-            .render({
-                'object': self.object,
-                'link': link,
-                'link_domain': link_domain,
-            })
-        body_html = get_template('mailing/event_submission.html') \
-            .render({
-                'object': self.object,
-                'link': link,
-                'link_domain': link_domain,
-            })
-        return body_txt, body_html
-
-
-class EventSubmissionConfirm(LoginNotRequiredMixin, TemplateView):
-    """Display confirmation of received workshop submission."""
-    template_name = 'forms/event_submission_confirm.html'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = 'Thanks for your submission'
-        return context
-
-
 class AllEventSubmissions(OnlyForAdminsMixin, FilteredListView):
     context_object_name = 'submissions'
     template_name = 'workshops/all_eventsubmissions.html'
@@ -2906,61 +2728,6 @@ def eventsubmission_assign(request, submission_id, person_id=None):
     submission = get_object_or_404(EventSubmissionModel, pk=submission_id)
     assign(request, submission, person_id)
     return redirect(submission.get_absolute_url())
-
-
-class DCSelfOrganizedEventRequest(LoginNotRequiredMixin, EmailSendMixin,
-                                  CreateViewContext):
-    "Display form for requesting self-organized workshops for Data Carpentry."
-    model = DCSelfOrganizedEventRequestModel
-    form_class = DCSelfOrganizedEventRequestForm
-    # we're reusing DC templates for normal workshop requests
-    template_name = 'forms/workshop_dc_request.html'
-    success_url = reverse_lazy('dc_workshop_selforganized_request_confirm')
-    email_fail_silently = False
-    email_kwargs = {
-        'to': settings.REQUEST_NOTIFICATIONS_RECIPIENTS,
-    }
-
-    def get_success_message(self, *args, **kwargs):
-        """Don't display a success message."""
-        return ''
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = 'Register a self-organized Data Carpentry workshop'
-        return context
-
-    def get_subject(self):
-        return ('DC: new self-organized workshop request from {} @ {}'
-                .format(self.object.name, self.object.organization))
-
-    def get_body(self):
-        link = self.object.get_absolute_url()
-        link_domain = settings.SITE_URL
-        body_txt = get_template('mailing/dc_self_organized.txt') \
-            .render({
-                'object': self.object,
-                'link': link,
-                'link_domain': link_domain,
-            })
-        body_html = get_template('mailing/dc_self_organized.html') \
-            .render({
-                'object': self.object,
-                'link': link,
-                'link_domain': link_domain,
-            })
-        return body_txt, body_html
-
-
-class DCSelfOrganizedEventRequestConfirm(LoginNotRequiredMixin, TemplateView):
-    """Display confirmation of a received self-organized workshop request."""
-    # we're reusing DC templates for normal workshop requests
-    template_name = 'forms/workshop_dc_request_confirm.html'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['title'] = 'Thanks for your submission'
-        return context
 
 
 class AllDCSelfOrganizedEventRequests(OnlyForAdminsMixin, FilteredListView):
@@ -3221,38 +2988,6 @@ def duplicates(request):
     }
 
     return render(request, 'workshops/duplicates.html', context)
-
-
-@login_not_required
-def trainingrequest_create(request):
-    """A form to let all users (no login required) to request Instructor
-    Training."""
-
-    form = TrainingRequestForm()
-    page_title = 'Apply for Instructor Training'
-
-    if request.method == 'POST':
-        form = TrainingRequestForm(request.POST)
-
-        if form.is_valid():
-            form.save()
-
-            # TODO: email notification?
-
-            context = {
-                'title': 'Thank you for applying for an instructor training.',
-            }
-            return render(request,
-                          'forms/trainingrequest_confirm.html',
-                          context)
-        else:
-            messages.error(request, 'Fix errors below.')
-
-    context = {
-        'title': page_title,
-        'form': form,
-    }
-    return render(request, 'forms/trainingrequest.html', context)
 
 
 @admin_required
